# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FinanceFlow API",
        "description": "FinanceFlow API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server development (localhost)"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server development (IP)"
        }
    ],
    "paths": {
...
        "/api/v1/reports": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get aggregated transaction report with optional filters",
                "description": "Returns income/expense summary totals and a breakdown by category. All filters are optional.",
                "operationId": "064f34c8712515671f265af3b0cb9096",
                "parameters": [
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-01"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-12-31"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "expense",
                            "enum": [
                                "income",
                                "expense",
                                "transfer"
                            ]
                        }
                    },
                    {
                        "name": "currency_code",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report data with summary and category breakdown",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "summary": {
                                                    "properties": {
                                                        "income": {
                                                            "type": "number",
                                                            "example": 3000
                                                        },
                                                        "expense": {
                                                            "type": "number",
                                                            "example": 1200
                                                        },
                                                        "net": {
                                                            "type": "number",
                                                            "example": 1800
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "by_category": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "filters": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reports/export": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Export report — coming in Stage 10 (PDF / Excel / CSV via queue)",
                "operationId": "d390891d76db63b7d6ca307d77745937",
                "responses": {
                    "501": {
                        "description": "Not yet implemented",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Export feature is planned for Stage 10. Will support PDF, Excel, and CSV formats."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}